confimação e  validação de exclusão

+++ app/control/DespesaFixa/DespesaList.php
@@ -14,6 +14,7 @@ use Adianti\Widget\Datagrid\TDataGridAction;
 use Adianti\Widget\Datagrid\TDataGridColumn;
 use Adianti\Widget\Datagrid\TPageNavigation;
 use Adianti\Widget\Dialog\TMessage;
+use Adianti\Widget\Dialog\TQuestion;
 use Adianti\Widget\Form\TCombo;
 use Adianti\Widget\Form\TEntry;
 use Adianti\Widget\Form\TLabel;
@@ -34,8 +35,9 @@ class DespesaList extends TPage
     private $deleteButton;
 
     use Adianti\base\AdiantiStandardListTrait;
-    
-    public function __construct(){
+
+    public function __construct()
+    {
 
         parent::__construct();
 
@@ -59,25 +61,25 @@ class DespesaList extends TPage
         $campo1 =  new TDBUniqueSearch('cpf', 'sample', 'FichaCadastral', 'cpf', 'cpf');
         $campo2         = new TDBUniqueSearch('anoMes', 'sample', 'AnoMes', 'descricao', 'descricao');
 
-        $this->form->addFields( [new TLabel('CPF')], [ $campo1 ]  );
-        $this->form->addFields( [new TLabel('Ano Mes')], [ $campo2 ]  );
+        $this->form->addFields([new TLabel('CPF')], [$campo1]);
+        $this->form->addFields([new TLabel('Ano Mes')], [$campo2]);
 
 
-          //Tamanho dos fields
-          $campo1->setSize('100%');
-          $campo1->setMinLength(0);
-          $campo2->setMinLength(0);
+        //Tamanho dos fields
+        $campo1->setSize('100%');
+        $campo1->setMinLength(0);
+        $campo2->setMinLength(0);
 
-        $this->form->setData( TSession::getValue( __CLASS__.'_filter_data') );
+        $this->form->setData(TSession::getValue(__CLASS__ . '_filter_data'));
 
         //Adicionar field de busca
         $btn = $this->form->addAction(_t('Find'), new TAction([$this, 'onSearch']), 'fa:search');
         $btn->class = 'btn btn-sm btn-primary';
-       // $this->form->addActionLink(_t('New'), new TAction(['DespesaForm', 'onEdit'], ['register_state' => 'false']), 'fa:plus green'  );
-        
-       $this->form->addActionLink(_t('New'), new TAction(['DespesaView', 'onEdit'], ['register_state' => 'false']), 'fa:plus green'  );
-       
-   
+        // $this->form->addActionLink(_t('New'), new TAction(['DespesaForm', 'onEdit'], ['register_state' => 'false']), 'fa:plus green'  );
+
+        $this->form->addActionLink(_t('New'), new TAction(['DespesaView', 'onEdit'], ['register_state' => 'false']), 'fa:plus green');
+
+
 
         //Criando a data grid
         $this->datagrid = new BootstrapDatagridWrapper(new TDataGrid);
@@ -85,11 +87,11 @@ class DespesaList extends TPage
 
         //Criando colunas da datagrid
         $column_1 = new TDataGridColumn('id', 'Codigo', 'left');
-        $column_2 = new TDataGridColumn('cpf', 'CPF', 'left', );
-        $column_2 = new TDataGridColumn('folha->descricao', 'Tipo', 'left', );
-        $column_3 = new TDataGridColumn('anoMes', 'Mês', 'left', );
-        $column_4 = new TDataGridColumn('vl_despesa', 'Despesas', 'left', );
-        $column_5 = new TDataGridColumn('saldo', 'Saldo', 'left', );
+        $column_2 = new TDataGridColumn('cpf', 'CPF', 'left',);
+        $column_2 = new TDataGridColumn('folha->descricao', 'Tipo', 'left',);
+        $column_3 = new TDataGridColumn('anoMes', 'Mês', 'left',);
+        $column_4 = new TDataGridColumn('vl_despesa', 'Despesas', 'left',);
+        $column_5 = new TDataGridColumn('saldo', 'Saldo', 'left',);
 
         $formato_vl_despesa = function ($value) {
             if (is_numeric($value)) {
@@ -106,7 +108,7 @@ class DespesaList extends TPage
             return $value;
         };
         $column_5->setTransformer($formato_vl_saldo);
-     
+
 
         //add coluna da datagrid
         $this->datagrid->addColumn($column_1);
@@ -116,17 +118,17 @@ class DespesaList extends TPage
         $this->datagrid->addColumn($column_5);
 
         //Criando ações para o datagrid
-        $column_1->setAction(new TAction([$this, 'onReload']), ['order'=> 'id']);
-        $column_2->setAction(new TAction([$this, 'onReload']), ['order'=> 'cpf']);
+        $column_1->setAction(new TAction([$this, 'onReload']), ['order' => 'id']);
+        $column_2->setAction(new TAction([$this, 'onReload']), ['order' => 'cpf']);
 
-        $action1 = new TDataGridAction(['DespesaView', 'onEdit'], ['id'=> '{id}', 'register_state' => 'false']);
-        $action2 = new TDataGridAction([ $this, 'onDelete'], ['id'=> '{id}']);
-        $action3 = new TDataGridAction([ 'DespesaMap', 'onReload'], ['id'=> '{id}']);
+        $action1 = new TDataGridAction(['DespesaView', 'onEdit'], ['id' => '{id}', 'register_state' => 'false']);
+        $action2 = new TDataGridAction([$this, 'onDelete'], ['id' => '{id}']);
+        $action3 = new TDataGridAction(['DespesaMap', 'onReload'], ['id' => '{id}']);
 
         //Adicionando a ação na tela
-        $this->datagrid->addAction($action1, _t('Edit'), 'fa:edit blue' );
-        $this->datagrid->addAction($action2, _t('Delete'), 'fa:trash-alt red' );
-        $this->datagrid->addAction($action3, 'Mapa', 'fa:light fa-map green' );
+        $this->datagrid->addAction($action1, _t('Edit'), 'fa:edit blue');
+        $this->datagrid->addAction($action2, _t('Delete'), 'fa:trash-alt red');
+        $this->datagrid->addAction($action3, 'Mapa', 'fa:light fa-map green');
 
 
         //Criar datagrid 
@@ -136,20 +138,20 @@ class DespesaList extends TPage
         $this->pageNavigation = new TPageNavigation;
         $this->pageNavigation->setAction(new TAction([$this, 'onReload']));
 
-      
+
 
         //Enviar para tela
         $panel = new TPanelGroup('', 'white');
         $panel->add($this->datagrid);
         $panel->addFooter($this->pageNavigation);
 
-          //Exportar
-          $drodown = new TDropDown('Exportar', 'fa:list');
-          $drodown->setPullSide('right');
-          $drodown->setButtonClass('btn btn-default waves-effect dropdown-toggle');
-          $drodown->addAction('Salvar como CSV', new TAction([$this, 'onExportCSV'], ['register_state' => 'false', 'static'=>'1']), 'fa:table green');
-          $drodown->addAction('Salvar como PDF', new TAction([$this, 'onExportPDF'], ['register_state' => 'false',  'static'=>'1']), 'fa:file-pdf red');
-          $panel->addHeaderWidget( $drodown);
+        //Exportar
+        $drodown = new TDropDown('Exportar', 'fa:list');
+        $drodown->setPullSide('right');
+        $drodown->setButtonClass('btn btn-default waves-effect dropdown-toggle');
+        $drodown->addAction('Salvar como CSV', new TAction([$this, 'onExportCSV'], ['register_state' => 'false', 'static' => '1']), 'fa:table green');
+        $drodown->addAction('Salvar como PDF', new TAction([$this, 'onExportPDF'], ['register_state' => 'false',  'static' => '1']), 'fa:file-pdf red');
+        $panel->addHeaderWidget($drodown);
 
         //Vertical container
         $container = new TVBox;
@@ -157,13 +159,10 @@ class DespesaList extends TPage
         $container->add(new TXMLBreadCrumb('menu.xml', __CLASS__));
         $container->add($this->form);
         $container->add($panel);
-    
+
         parent::add($container);
-
-       
-
     }
-   
+
     public function onDelete($param)
     {
         try {
@@ -171,22 +170,41 @@ class DespesaList extends TPage
 
                 $id = $param['key'];
 
-                TTransaction::open('sample');
-                $despesa = new Despesa($id);
-                ItemDespesa::where('despesa_id', '=', $despesa->id)->delete();
-                $despesa->delete();
-                
+
+                // Exiba uma mensagem de confirmação antes de excluir
+                $action = new TAction(array($this, 'delete'));
+                $action->setParameter('id', $id);
+
+                new TQuestion('Deseja excluir?', $action);
             }
-            new TMessage('info', 'Registo Excluido', $this->afterSaveAction); //$this->afterSaveAction
 
             TTransaction::close();
         } catch (Exception $e) {
             new TMessage('error', $e->getMessage(), $this->afterSaveAction);
             TTransaction::rollback();
         }
-    
-  
-}
+    }
 
-  
+
+    public function delete($param)
+    {
+        try {
+            TTransaction::open('sample');
+
+            if (isset($param['id'])) {
+                $id = $param['id'];
+
+                $despesa = new Despesa($id);
+                ItemDespesa::where('despesa_id', '=', $despesa->id)->delete();
+                $despesa->delete();
+                new TMessage('info', 'Registo Excluido', $this->afterSaveAction); //$this->afterSaveAction
+                $this->onReload([]);
+            }
+
+            TTransaction::close();
+        } catch (Exception $e) {
+            new TMessage('error', $e->getMessage(), $this->afterSaveAction);
+            TTransaction::rollback();
+        }
+    }
 }
